Update the main event loop
The event loop was using stdio based functions which because of the
internal buffering do not work as well as just doing everything with
low level select/read loops. This improvement should let the plugin
process events faster so that auditd can move on to other things.

The auparse_set_eoe_timeout function was added in 3.0.1, so the dependencies
were updated. If compatibility with older audit libraries is needed, delete
that function and 2.5.1 should be good enough to run the new event loop.